remove alert PromfetcherTooManyFailureDuringFetch
This alert trigger when `promfetcher` is unable to access the `/metrics` of any application, which might, in some cases, the problem of the App and not promfetcher.

We suggest users to either reimplement this alert on their side or to monitor a specific application to test wenether promfetcher might be failing or not.